+++ gpxity/test/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+# Copyright (c) 2019 Wolfgang Rohdewald <wolfgang@rohdewald.de>
+# See LICENSE for details.
+
+
+"""define things that should be visible to the user."""